nice

+++ 0x07-pointers_arrays_strings/0-memset.c
@@ -1,17 +1,16 @@
 #include "main.h"
-#include <stdio.h>
 
 /**
  * _puts_recursion - print a string followd by new line
  * @s: contains a string
  * @b: contains a charachter
  * @n: contains a unsigned integer
- * Return: void.
+ * Return: a pointer to memory area.
  */
 
 char *_memset(char *s, char b, unsigned int n)
 {
-	for (int i = 0; i < n; i++)
+	for (int i = 0; i < (int)n; i++)
 	{
 		s[i] = b;
 	}
